Refactor TableSkeleton
