--- apps/balancer-tools/src/app/apr/(components)/(skeleton)/PoolTableSkeleton.tsx
@@ -1,88 +1,0 @@
-export default function PoolTableSkeleton() {
-  return (
-    <div
-      role="status"
-      className="w-full animate-pulse space-y-4 rounded border p-4 shadow md:p-6 border-blue6 bg-blue3"
-    >
-      <div className="flex items-center justify-between border-b border-blue6 pb-4">
-        <div className="flex gap-3">
-          <div className="h-2.5 w-12 rounded-full bg-blue6"></div>
-          <div className="h-2.5 w-12 rounded-full bg-blue6"></div>
-        </div>
-        <div className="flex gap-3">
-          <div className="h-2.5 w-12 rounded-full bg-blue6"></div>
-          <div className="h-2.5 w-12 rounded-full bg-blue6"></div>
-          <div className="h-2.5 w-12 rounded-full bg-blue6"></div>
-        </div>
-      </div>
-      <div className="flex items-center justify-between">
-        <div className="flex gap-3">
-          <div className="h-2.5 w-12 rounded-full bg-blue6"></div>
-          <div className="h-2.5 w-12 rounded-full bg-blue6"></div>
-        </div>
-        <div className="flex gap-3">
-          <div className="h-2.5 w-12 rounded-full bg-blue6"></div>
-          <div className="h-2.5 w-12 rounded-full bg-blue6"></div>
-          <div className="h-2.5 w-12 rounded-full bg-blue6"></div>
-        </div>
-      </div>
-      <div className="flex items-center justify-between pt-4">
-        <div className="flex gap-3">
-          <div className="h-2.5 w-12 rounded-full bg-blue6"></div>
-          <div className="h-2.5 w-12 rounded-full bg-blue6"></div>
-        </div>
-        <div className="flex gap-3">
-          <div className="h-2.5 w-12 rounded-full bg-blue6"></div>
-          <div className="h-2.5 w-12 rounded-full bg-blue6"></div>
-          <div className="h-2.5 w-12 rounded-full bg-blue6"></div>
-        </div>
-      </div>
-      <div className="flex items-center justify-between pt-4">
-        <div className="flex gap-3">
-          <div className="h-2.5 w-12 rounded-full bg-blue6"></div>
-          <div className="h-2.5 w-12 rounded-full bg-blue6"></div>
-        </div>
-        <div className="flex gap-3">
-          <div className="h-2.5 w-12 rounded-full bg-blue6"></div>
-          <div className="h-2.5 w-12 rounded-full bg-blue6"></div>
-          <div className="h-2.5 w-12 rounded-full bg-blue6"></div>
-        </div>
-      </div>
-      <div className="flex items-center justify-between pt-4">
-        <div className="flex gap-3">
-          <div className="h-2.5 w-12 rounded-full bg-blue6"></div>
-          <div className="h-2.5 w-12 rounded-full bg-blue6"></div>
-        </div>
-        <div className="flex gap-3">
-          <div className="h-2.5 w-12 rounded-full bg-blue6"></div>
-          <div className="h-2.5 w-12 rounded-full bg-blue6"></div>
-          <div className="h-2.5 w-12 rounded-full bg-blue6"></div>
-        </div>
-      </div>
-      <div className="flex items-center justify-between pt-4">
-        <div className="flex gap-3">
-          <div className="h-2.5 w-12 rounded-full bg-blue6"></div>
-          <div className="h-2.5 w-12 rounded-full bg-blue6"></div>
-        </div>
-        <div className="flex gap-3">
-          <div className="h-2.5 w-12 rounded-full bg-blue6"></div>
-          <div className="h-2.5 w-12 rounded-full bg-blue6"></div>
-          <div className="h-2.5 w-12 rounded-full bg-blue6"></div>
-        </div>
-      </div>
-      <div className="flex items-center justify-between pt-4">
-        <div className="flex gap-3">
-          <div className="h-2.5 w-12 rounded-full bg-blue6"></div>
-          <div className="h-2.5 w-12 rounded-full bg-blue6"></div>
-        </div>
-        <div className="flex gap-3">
-          <div className="h-2.5 w-12 rounded-full bg-blue6"></div>
-          <div className="h-2.5 w-12 rounded-full bg-blue6"></div>
-          <div className="h-2.5 w-12 rounded-full bg-blue6"></div>
-        </div>
-      </div>
-
-      <span className="sr-only">Loading...</span>
-    </div>
-  );
-}
