+++ apps/balancer-tools/src/app/apr/round/[roundId]/page.tsx
@@ -2,7 +2,7 @@ import { Suspense } from "react";
 
 import ChartSkelton from "../../(components)/(skeleton)/ChartSkelton";
 import KpisSkeleton from "../../(components)/(skeleton)/KpisSkeleton";
-import PoolTableSkeleton from "../../(components)/(skeleton)/PoolTableSkeleton";
+import TableSkeleton from "../../(components)/(skeleton)/TableSkeleton";
 import getFilteredRoundApiUrl from "../../(utils)/getFilteredApiUrl";
 import Breadcrumb from "../(components)/Breadcrumb";
 import PoolTableWrapper from "../(components)/PoolTableWrapper";
@@ -35,7 +35,7 @@ export default function Page({
       <Suspense fallback={<ChartSkelton />}>
         <TopPoolsChartWrapper roundId={roundId} />
       </Suspense>
-      <Suspense fallback={<PoolTableSkeleton />}>
+      <Suspense fallback={<TableSkeleton colNumbers={10} />}>
         <PoolTableWrapper
           roundId={roundId}
           url={getFilteredRoundApiUrl(searchParams, roundId)}
